requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7
alembic==1.12.0
pydantic==2.3.0
flasgger==0.9.7.1
gunicorn==21.2.0
uvicorn==0.23.1
asgiref==3.8.0
